add schemas and database files
