+++ database/BPower.js
@@ -1,0 +1,13 @@
+const mongoose = require('mongoose');
+const db = require('./index.js');
+mongoose.Promise = global.Promise;
+
+let buyingPowerSchema = new mongoose.Schema({
+  id: Number,
+  user: String, 
+  marketPrice: Number,
+});
+
+const BuyingPower = mongoose.model('Buying Power', buyingPowerSchema);
+
+module.exports = BuyingPower;
