+++ package.json
@@ -7,7 +7,8 @@
     "build": "webpack --mode production",
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "webpack -d --watch",
-    "start": "nodemon server/index.js"
+    "start": "nodemon server/index.js",
+    "seed": "node database/seed.js"
   },
   "repository": {
     "type": "git",
